refactor: startTracker calls startWebSocketServer directly
