+++ packages/network/test/integration/killing-dead-connections.test.ts
@@ -1,10 +1,11 @@
 /* eslint-disable no-underscore-dangle */
 import { waitForEvent } from 'streamr-test-utils'
 
-import { startServerWsEndpoint, ServerWsEndpoint } from '../../src/connection/ServerWsEndpoint'
+import { ServerWsEndpoint } from '../../src/connection/ServerWsEndpoint'
 import { PeerInfo } from '../../src/connection/PeerInfo'
 import { ClientWsEndpoint } from '../../src/connection/ClientWsEndpoint'
 import { DisconnectionReason, Event } from "../../src/connection/AbstractWsEndpoint"
+import { startServerWsEndpoint } from '../utils'
 
 const STATE_OPEN = 1
