+++ packages/network/test/integration/ws-endpoint.test.ts
@@ -2,11 +2,12 @@ import { Tracker } from '../../src/logic/Tracker'
 import WebSocket from 'ws'
 import { waitForEvent, wait } from 'streamr-test-utils'
 
-import { startServerWsEndpoint, ServerWsEndpoint } from '../../src/connection/ServerWsEndpoint'
+import { ServerWsEndpoint } from '../../src/connection/ServerWsEndpoint'
 import { PeerInfo } from '../../src/connection/PeerInfo'
 import { startTracker } from '../../src/composition'
 import { ClientWsEndpoint } from '../../src/connection/ClientWsEndpoint'
 import { DisconnectionCode, Event } from "../../src/connection/AbstractWsEndpoint"
+import { startServerWsEndpoint } from '../utils'
 
 describe('ws-endpoint', () => {
     const endpoints: ServerWsEndpoint[] = []
